+++ src/util/LineChart/index.js
@@ -1,4 +1,4 @@
-import { select, scaleTime, scaleLinear, axisLeft, axisBottom, extent, line, area } from 'd3';
+import { select, scaleTime, scaleLinear, axisLeft, axisBottom, extent, line, area, max } from 'd3';
 import { initSvg, resize } from './linechart-helpers';
 
 export default class LineChart {
@@ -30,6 +30,8 @@ export default class LineChart {
     this.period = period;
     this.obj = obj;
 
+    const highestValue = max(progressionList, (d) => d.value);
+
     this.width = this.svg.node().getBoundingClientRect().width;
     resize.call(this);
 
@@ -38,7 +40,12 @@ export default class LineChart {
     const endDate = period.endDate && period.endDate.toDate ? period.endDate.toDate() : new Date(period.endDate);
 
     this.x.domain([startDate, endDate]);
-    this.y.domain(extent([obj.startValue, obj.targetValue]));
+    this.y.domain(
+      extent([
+        obj.startValue,
+        obj.startValue < obj.targetValue && highestValue > obj.targetValue ? highestValue : obj.targetValue,
+      ])
+    );
 
     this.yAxis.transition().call(axisLeft(this.y));
     this.xAxis.transition().call(axisBottom(this.x).ticks(4));